Initial commit
Includes a barely working version of CChess. Move generation is completely finished and is somewhat efficient when used with multi-threading and hashing. The search algorithm itself is very poor however, and can only search up to a depth of roughly 4 without taking forever. Even then, it still blunders pieces sometimes and is generally not very good. The biggest improvement right now would probably be improving memory management during the search and adding move ordering to quiescence search. There is also a possible bug with the hash table that results in different moves played in the same position. Another thing to consider is that quiescence search might not be as stable for even depths.